requirements.txt
discord.py[voice]
requests
disrank
python-dotenv
g4f
curl_cffi
python-dateutil
phonenumbers
email_validator
anyio
pytest
pillow
demoparser2